Remove the re-try logic from svn_fs_fs__path_rev_absolute().
The re-try logic was introduced in r875097 and reduced but did not eliminate
the window of opportunity for the caller to use an out-of-date result.
Since r1040832, all its callers correctly account for the possibility of an
out-of-date result due to a concurrent packing operation.

See the email thread <http://svn.haxx.se/dev/archive-2010-12/0019.shtml>,
subject "Re: svn commit: r1040832 - Port a fix for a FSFS packing race".

* subversion/libsvn_fs_fs/fs_fs.c
  (svn_fs_fs__path_rev_absolute): Remove the re-try logic.

* subversion/libsvn_fs_fs/fs_fs.h
  (svn_fs_fs__path_rev_absolute): Update the doc string accordingly.


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@1043408 13f79535-47bb-0310-9956-ffa450edef68