Make wc_db's base_get_info() report information about whether a node should only be handled as an update root. (Currently only true for file externals).
While I still think we need to clean up the file externals implementation
sometime in the future, we'll have to ship with file externals in 1.7. This
change will allow reducing the number of SQLite queries by at least the
number of nodes in the working copy. And as this is a library private api,
we can always remove this support in a future version.

A separate commit will add this boolean to svn_wc__db_base_get_info() and
update it's +- 50 callers.

* subversion/libsvn_wc/wc-queries.sql
  (STMT_SELECT_BASE_NODE,
   STMT_SELECT_BASE_NODE_WITH_LOCK): Insert update_root value at position 14.

* subversion/libsvn_wc/wc_db.c
  (base_get_info): Add update_root argument.
  (get_info_for_copy, bump_node, bump_nodes, bump_revisions_post_commit,
   svn_wc__db_temp_op_set_base_incomplete): Update callers.


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@1081753 13f79535-47bb-0310-9956-ffa450edef68