[ Note from the future: reverted in r1088409 ]
Boost peek full text cache hit rates from approx. 50% to 100%.

In many cases, we don't know the full text size in advance
making it impossible to detect the end of text in rep_read_contents
and to put it into the cache. Therefore, store all other full texts
upon closing the respective stream. 

Note that it is still useful to store the item content asap since
we have no control (here) on the pool life time. Therefore, the
item might not get cached for a very long time.

* subversion/libsvn_fs_fs/fs_fs.c
  (cache_rep): new utility function that caches the baton's current full text.
  (rep_read_contents, rep_read_contents_close): call the new function

git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@1088382 13f79535-47bb-0310-9956-ffa450edef68