Following up on r1089903, fix issue #3334 in a better way.
After this patch the make_copy operation just introduces a copy layer below,
a possible working layer. This fixes cases where deletes broke the tree and
fixes the issue #3334 test that verified that a direct copy and a copy made
via make_copy behave in the same way.

This patch fixes the behavior for these tests, but doesn't produce a proper
mixed revision copy yet.

* subversion/libsvn_wc/wc_db.c
  (make_copy_txn): Update schema. Fetch op_depth and only apply deleted
    transformations when we are changing existing rows.
    (Currently this happens never, but it will when we create a proper mixed
     revision copy).

* subversion/tests/cmdline/update_tests.py
  (tree_conflict_uc2_schedule_re_add): Remove XFail marking.

* subversion/tests/libsvn_wc/op-depth-test.c
  (test_temp_op_make_copy): Update expected results to assume higher layers
    don't change when a copy is added below them.


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@1090053 13f79535-47bb-0310-9956-ffa450edef68