Separate the implementations of the working copy diff editor (svn diff -r 123 PATH and svn diff -r 123:BASE) and the local diff (svn diff PATH).
These two diffs currently share a lot of code, but they have to look at
restructuring changes in a different way.

This code was originally designed before Subversion 1.4, where all
restructuring changes where simply handled by the repository side of the diff:
we reported the original checked out version as deleted in the adm crawler
and received the restructurings from the repository.

With the invention of the revert base (in 1.4) and now with full local layering
this model grows apart from the local diff, which just reports changes against
the origin of the file. Separating the implementation makes things much
simpler.

No functional changes (yet).

* subversion/libsvn_wc/diff.c
  Deleted file.

* subversion/libsvn_wc/diff_editor.c
  New file; copied from subversion/libsvn_wc/diff.c
  (header): Updated.
  (svn_wc_diff6): Remove function.

* subversion/libsvn_wc/diff_local.c
  New file; copied from subversion/libsvn_wc/diff.c
  (*): Remove all functions and structs only used from the diff editor.


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@1097718 13f79535-47bb-0310-9956-ffa450edef68