Ok, thinking this over at 4 AM. I don't think we are willing to accept the performance regression of r1101473 in svn revert -R.  So I accept that we don't have to revert that are not strictly as checked out, but are still unmodified according to status.
In most cases this reverts to the behavior of before r1101473, as
before that we accepted recorded information as 'the file is not
modified'. But I'm introducing the change now that when the file might
be modified, we do the less strict comparison.

The only way to have both the performance and the strict comparison
would be to store another boolean with the recorded information to
tell to what level we performed the recorded comparison.  (And should
we then cache the strict or the not strict?)

[Note from the future: This reverts 'svn revert' to the 1.6.x behavior
 of using the recorded information as hint.  But this breaks one of
 the revert tests because that tests assumes that 'svn revert' notices
 keyword expansion changes.  More details in r1101730.]

* subversion/libsvn_wc/adm_ops.c
  (revert_restore): No longer ask svn_wc__internal_file_modified_p()
    to perform an exact comparison.

git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@1101728 13f79535-47bb-0310-9956-ffa450edef68