Implement some (initially very simplistic) verification of the WC metadata in the DB.  When compiled in SVN_DEBUG mode, make 'svn cleanup' run these checks and report any violations.
The purpose of this is to get some verification code started and for
developers to be able to run it and extend it.  Running it in 'svn cleanup'
is just an initial and rather arbitrary UI choice; we could instead use a
new subcommand or option or an external program.

* subversion/libsvn_wc/cleanup.c
  (cleanup_internal): Call svn_wc__db_verify() if SVN_DEBUG is defined.

* subversion/libsvn_wc/wc_db.c
  (VERIFY): New macro.
  (db_verify, svn_wc__db_verify): New functions.

* subversion/libsvn_wc/wc_db.h
  (svn_wc__db_verify): New function.

* subversion/libsvn_wc/wc-queries.sql
  (STMT_SELECT_ALL_NODES): New query.


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@1101848 13f79535-47bb-0310-9956-ffa450edef68