Write the rest of the code for redirecting content to the PENDING structures when the parser is 'paused'.
This code is not yet enabled because the EOF handling needs to be
corrected. Right now, when the network content is finished, this code
free's the XML parser. But we may need to keep it around for parsing
the pending content.

And I just realized that the network code cannot simply inject the
arriving content into the parser. There may be stuff in PENDING that needs
to be inject first. That is, this code may inject it out of order.

Committing anyway as this demonstrates more of the direction of this code.

* subversion/libsvn_ra_serf/util.c:
  (svn_ra_serf__handle_xml_parser): pre-position the spill file to append
    content in write_to_pending(). switch to the PARSE_CHUNK_SIZE symbol
    rather than a bare constant. add logic to alternately write to the
    PENDING structures, or inject it into the xml parser.


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@1138119 13f79535-47bb-0310-9956-ffa450edef68